+++ types/src/block/output/nft_id.rs
@@ -8,15 +8,15 @@ impl_id!(pub NftId, 32, "TODO.");
 #[cfg(feature = "serde")]
 string_serde_impl!(NftId);
 
-impl From<OutputId> for NftId {
-    fn from(output_id: OutputId) -> Self {
+impl From<&OutputId> for NftId {
+    fn from(output_id: &OutputId) -> Self {
         Self::from(output_id.hash())
     }
 }
 
 impl NftId {
     ///
-    pub fn or_from_output_id(self, output_id: OutputId) -> Self {
+    pub fn or_from_output_id(self, output_id: &OutputId) -> Self {
         if self.is_null() { Self::from(output_id) } else { self }
     }
 }